Fix assign moderation core import for deployment
